add controller and error handler for author entity + update entities

+++ src/main/java/com/younes/reskilingproject/bookManagement/service/AuthorService/ImplAuthorService.java
@@ -1,6 +1,7 @@
 package com.younes.reskilingproject.bookManagement.service.AuthorService;
 
 import com.younes.reskilingproject.bookManagement.entity.bookStore.Author;
+import com.younes.reskilingproject.bookManagement.errorHandler.authorError.AuthorNotFoundException;
 import com.younes.reskilingproject.bookManagement.repository.AuthorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,7 +17,7 @@ public class ImplAuthorService implements AuthorService {
         this.authorRepository = authorRepository;
     }
     @Override
-    public Author addAuthor(Author newAuthor) {
+    public Author saveAuthor(Author newAuthor) {
         return authorRepository.save(newAuthor);
     }
     @Override
@@ -25,10 +26,10 @@ public class ImplAuthorService implements AuthorService {
     }
     @Override
     public Author findAuthor(long id) {
-        return authorRepository.findById(id).orElse(null);
+        return authorRepository.findById(id).orElseThrow(() -> new AuthorNotFoundException("Author Not found"));
     }
     @Override
     public void deleteAuthor(long id) {
-        authorRepository.deleteById(id);
+        if(findAuthor(id) != null) authorRepository.deleteById(id);
     }
 }
